General code clean-up and file refactoring
