+++ Library/UniversalBeacon.Library.Core/Interop/BLEManufacturerData.cs
@@ -1,59 +1,27 @@
-﻿using System;
+﻿// Copyright 2015 - 2017 Chris Tacke. All rights reserved. 
+// https://github.com/andijakl/universal-beacon 
+// 
+// Based on the Eddystone specification by Google, 
+// available under Apache License, Version 2.0 from
+// https://github.com/google/eddystone
+// 
+// Licensed under the Apache License, Version 2.0 (the "License"); 
+// you may not use this file except in compliance with the License. 
+// You may obtain a copy of the License at 
+// 
+//    http://www.apache.org/licenses/LICENSE-2.0 
+// 
+// Unless required by applicable law or agreed to in writing, software 
+// distributed under the License is distributed on an "AS IS" BASIS, 
+// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied. 
+// See the License for the specific language governing permissions and 
+// limitations under the License. using System;
+using System;
 using System.Collections.Generic;
 using System.Text;
 
 namespace UniversalBeaconLibrary
 {
-    // https://www.bluetooth.com/specifications/assigned-numbers/16-bit-uuids-for-members
-    public enum BTMember
-    {
-        Estimote = 0xFE9A,
-        GoogleEddystone = 0xFEAA
-    }
-
-    // https://www.bluetooth.com/specifications/assigned-numbers/company-identifiers
-    public enum BTCompany
-    {
-    }
-
-    // https://www.bluetooth.com/specifications/assigned-numbers/generic-access-profile
-    public enum BLEPacketType
-    {
-        Invalid = 0x00,
-        Flags = 0x01,
-        LocalName = 0x09,
-        UUID16List = 0x03,
-        ServiceData = 0x16,
-        ManufacturerData = 0xff
-    }
-
-    public class FlagData : BLERecord
-    {
-        public byte Flags { get; private set; }
-
-        public FlagData(BLEPacketType packetType, byte flags)
-            : base(packetType)
-        {
-            Flags = flags;
-        }
-    }
-
-    public class NameData : BLERecord
-    {
-        private string m_name;
-
-        public NameData(BLEPacketType packetType, byte[] data)
-            : base(packetType)
-        {
-            m_name = Encoding.ASCII.GetString(data);
-        }
-
-        public string Name
-        {
-            get { return m_name; }
-        }
-    }
-
     public sealed class BLEManufacturerData : BLERecord
     {
         public ushort CompanyId { get; set; }
@@ -72,27 +40,5 @@ namespace UniversalBeaconLibrary
         {
         }
 
-    }
-
-
-    public class GenericRecord : BLERecord
-    {
-        public byte[] RawData { get; private set; }
-
-        public GenericRecord(BLEPacketType packetType, byte[] data)
-            : base(packetType)
-        {
-            RawData = data;
-        }
-    }
-
-    public abstract class BLERecord
-    {
-        public BLEPacketType PacketType { get; set; }
-
-        public BLERecord(BLEPacketType packetType)
-        {
-            PacketType = packetType;
-        }
     }
 }
